==> notebooks/1-population-data-cleaning.ipynb <==
# Data manipulation
import pandas as pd # data manipulation and dataframes

# Pandas configuration
pd.set_option('display.max_columns', None)  # shows all columns
pd.set_option('display.max_colwidth', None)  # shows all cell content

# Reads csv and creates a dataframe
airbnb_porto = pd.read_csv('../data/airbnb_porto.csv')
ine_porto = pd.read_csv('../raw/ine-portugal/ine21_district_porto.csv', delimiter=';')

airbnb_lisbon = pd.read_csv('../data/airbnb_lisbon.csv', encoding='utf-8')
ine_lisbon = pd.read_csv('../raw/ine-portugal/ine21_district_lisbon.csv', delimiter=';')

# Porto

# Cleans raw population data from portuguesh ine and gets a csv with the population for each district in Porto city

ine_porto[['District', 'Population']] = ine_porto[['Place of residence at Census date [2021] (NUTS - 2013)', 'S7A2021:2021-T:MF-T:Total']]
ine_porto = ine_porto[['District', 'Population']]

aux = ine_porto['District'].tolist()
dis = []

for a in aux:

    dis.append(a.split(':')[1])

ine_porto['District'] = dis
ine_porto.to_csv('../raw/eda-result/population_by_district_porto.csv', index = False, sep = ',')

# Renames district names in clean airbnb csv because they were written all in caps in raw airbnb data

hada_dict = {}
old_district_names = sorted(airbnb_porto['District'].unique().tolist())
new_district_names = ine_porto['District'].tolist()

for i in range(0, len(old_district_names)):

    hada_dict[old_district_names[i]] = new_district_names[i]

airbnb_porto['District'] = airbnb_porto['District'].map(hada_dict)

airbnb_porto.to_csv(f'../raw/eda-result/airbnb_porto.csv', index = False, sep = ',')

# Lisbon

ine_lisbon[['District', 'Population']] = ine_lisbon[['Place of residence at Census date [2021] (NUTS - 2013)', 'S7A2021:2021-T:MF-T:Total']]
ine_lisbon = ine_lisbon[['District', 'Population']]

aux = ine_lisbon['District'].tolist()
dis = []

for a in aux:

    dis.append(a.split(':')[1])

ine_lisbon['District'] = dis

ine_lisbon = ine_lisbon.sort_values(by=['District'])

ine_lisbon.to_csv(f'../raw/eda-result/population_by_district_lisbon.csv', index = False, sep = ',')

hada_dict = {}
old_district_names = sorted(airbnb_lisbon['District'].unique().tolist())
new_district_names = ine_lisbon['District'].tolist()

for i in range(0, len(old_district_names)):

    hada_dict[old_district_names[i]] = new_district_names[i]

airbnb_lisbon['District'] = airbnb_lisbon['District'].map(hada_dict)

airbnb_lisbon.to_csv(f'../raw/eda-result/airbnb_lisbon.csv', index = False, sep = ',')
