==> notebooks/2-testing-viz-on-px.ipynb <==
# Data manipulation
import pandas as pd # data manipulation and dataframes

# Data visualization
import plotly.express as px

# Pandas configuration
pd.set_option('display.max_columns', None)  # shows all columns
pd.set_option('display.max_colwidth', None)  # shows all cell content

# Reads csvs
df = pd.read_csv('../data/airbnb_madrid.csv', index_col = 0)

# Index
df.reset_index(drop = False, inplace = True)

# Color palette
blues = pd.read_csv('../data/blues.csv')

unique_district = df['District'].unique()
unique_neighbourhood = df['Neighbourhood'].unique()

airbnbs_per_district_madrid = {'district': [], 'total': []}

for u in unique_district:

    airbnbs_per_district_madrid['district'].append(u)
    airbnbs_per_district_madrid['total'].append(df[df['District'] == u].value_counts().sum())

viz1 = pd.DataFrame.from_dict(airbnbs_per_district_madrid)

viz1 = viz1.sort_values(by=['total'], ascending=False)
viz1

fig = px.histogram(df, x=viz1['district'], 
                   y=viz1['total'], 
                   color=viz1['district'], 
                   color_discrete_sequence=blues['Code'].tolist())

fig.update_xaxes(categoryorder = 'total descending')
fig.update_layout(xaxis_showticklabels=False)

fig.show()

airbnbs_per_neighbourhood_madrid = {'neighbourhood': [], 'total': []}
unique_neighbourhood = df['Neighbourhood'][df['District'] == 'Centro'].unique()

for u in unique_neighbourhood:

    airbnbs_per_neighbourhood_madrid['neighbourhood'].append(u)
    airbnbs_per_neighbourhood_madrid['total'].append(df[df['Neighbourhood'] == u].value_counts().sum())

viz2 = pd.DataFrame.from_dict(airbnbs_per_neighbourhood_madrid)
viz2

viz2
