==> notebooks/3-testing-kepler.ipynb <==
from keplergl import KeplerGl

# Data manipulation
import pandas as pd # data manipulation and dataframes

import pickle

import json

# Reads csvs
df_madrid = pd.read_csv('../data/airbnb_madrid.csv', index_col = 0)
df_madrid.reset_index(drop = False, inplace = True)

df_porto = pd.read_csv('../data/airbnb_porto.csv', index_col = 0)
df_porto.reset_index(drop = False, inplace = True)

df_lisbon = pd.read_csv('../data/airbnb_lisbon.csv', index_col = 0)
df_lisbon.reset_index(drop = False, inplace = True)

# GeoJSON as string
with open('../data/neighbourhood_madrid.geojson', 'r') as f:
    
    geojson1 = f.read()

# GeoJSON as string
with open('../data/neighbourhood_porto.geojson', 'r') as f:
    
    geojson2 = f.read()

# GeoJSON as string
with open('../data/neighbourhood_lisbon.geojson', 'r') as f:
    
    geojson3 = f.read()

map_4 = KeplerGl(show_docs=False)
map_4.add_data(data=df_madrid, name='airbnb_madrid')
map_4.add_data(data=geojson1, name='neighbourhood_madrid')
map_4

with open('config_madrid_rent.pickle', 'wb') as f:
    pickle.dump(map_4.config, f)

with open('config_mad_districts.pickle', 'rb') as configuration:
    config1 = pickle.load(configuration)

map_5 = KeplerGl(height=400, data={"airbnb_madrid": df_madrid[df_madrid['District'] == 'Salamanca']}, config=config1, show_docs=False);
map_5.add_data(data=geojson1, name='neighbourhood_madrid');
map_5


map_7 = KeplerGl(show_docs=False)
map_7.add_data(data=df_porto, name='airbnb_porto')
map_7.add_data(data=geojson2, name='neighbourhood_porto')
map_7

with open('config_porto_districts.pickle', 'wb') as f:
    pickle.dump(map_7.config, f)

map_8 = KeplerGl(show_docs=False)
map_8.add_data(data=df_lisbon, name='airbnb_lisbon')
map_8.add_data(data=geojson3, name='neighbourhood_lisbon')
map_8

with open('config_lisbon_districts.pickle', 'wb') as f:
    pickle.dump(map_8.config, f)
