==> notebooks/4-drafts.ipynb <==
# Data manipulation
import pandas as pd # data manipulation and dataframes

# Data visualization
import plotly.express as px
import matplotlib.pyplot as plt
import seaborn as sns

import calendar

# Pandas configuration
pd.set_option('display.max_columns', None)  # shows all columns
pd.set_option('display.max_colwidth', None)  # shows all cell content

# Reads csv
airbnb_madrid = pd.read_csv('../raw/eda-result/airbnb_madrid.csv', encoding='utf-8')
listings = pd.read_csv('../raw/airbnb/madrid/listings.csv', encoding='utf-8')

listings.head(1)

len(listings['name'].unique().tolist())

airbnb_madrid.head(1)

#airbnb_madrid['Last review'] = airbnb_madrid['Last review'].fillna('2023-07-26')
airbnb_madrid = airbnb_madrid.dropna()

airbnb_madrid["Last review"] = pd.to_datetime(airbnb_madrid["Last review"])
airbnb_madrid['Year'] = airbnb_madrid['Last review'].dt.strftime('%Y')


airbnb_madrid[airbnb_madrid['Last review'].dt.year == 2018].value_counts().sum()

airbnb_madrid['Last review'].dt.year.unique().tolist()

months = airbnb_madrid['Last review'].dt.month.unique().tolist()

airbnb_madrid['Last review'].dt.month[airbnb_madrid['Last review'].dt.year == 2022].value_counts().sum()

districts = airbnb_madrid['District'].unique().tolist()

total_airbnbs_by_month_2022 = {'Month': [], 'Airbnb reviews': []}

for m in months:
    total_airbnbs_by_month_2022['Month'].append(m)
    total_airbnbs_by_month_2022['Airbnb reviews'].append(airbnb_madrid[(airbnb_madrid['Last review'].dt.year == 2022) & (airbnb_madrid['Last review'].dt.month == m)].value_counts().sum())

total_airbnbs_by_month_2023 = {'Month': [], 'Airbnb reviews': []}

for m in months:
    total_airbnbs_by_month_2023['Month'].append(m)
    total_airbnbs_by_month_2023['Airbnb reviews'].append(airbnb_madrid[(airbnb_madrid['Last review'].dt.year == 2023) & (airbnb_madrid['Last review'].dt.month == m)].value_counts().sum())

total_airbnbs_by_month_2022

total_airbnbs_by_month_2022 = pd.DataFrame.from_dict(total_airbnbs_by_month_2022)
total_airbnbs_by_month_2022

total_airbnbs_by_month_2023

df2022 = pd.DataFrame.from_dict(total_airbnbs_by_month_2022)
df2022 = df2022.sort_values(by=['Month'])
df2022

df2023 = pd.DataFrame.from_dict(total_airbnbs_by_month_2023)
df2023 = df2023.sort_values(by=['Month'])
df2023

list_months_2022 = ['January 2022', 'February 2022', 'March 2022', 'April 2022', 'May 2022', 'June 2022', 'July 2022', 'August 2022', 'September 2022', 'October 2022', 'November 2022', 'December 2022']

df2022['Month'] = list_months_2022
df2022

list_months_2023 = ['January 2023', 'February 2023', 'March 2023', 'April 2023', 'May 2023', 'June 2023', 'July 2023', 'August 2023', 'September 2023', 'October 2023', 'November 2023', 'December 2023']

df2023['Month'] = list_months_2023
df2023

frames = [df2022, df2023]

result = pd.concat(frames)

result = result.loc[(result!=0).all(axis=1)]
result.to_csv('../raw/eda-result/reviews_per_month_airbnb.csv', index = False, sep = ',')
result


cmap_mpl = plt.get_cmap("YlOrRd")
cmap_mpl

fig = px.histogram(result, x=result['Month'], y=result['Airbnb reviews'], text_auto='.2s', color=result['Month'])
fig.update_traces(textfont_size=12, textangle=0, textposition='outside', cliponaxis=False)


fig.show()
