==> notebooks/6-drafts-copy-copy.ipynb <==
import pandas as pd
import plotly.express as px

# Pandas configuration
pd.set_option('display.max_columns', None)  # shows all columns
pd.set_option('display.max_colwidth', None)  # shows all cell content

# Reads csv
rent_madrid = pd.read_csv('../data/rent_price_evolution_by_district_madrid.csv', encoding='utf-8')
rent_madrid.head(1)

df = rent_madrid

df_aux = pd.DataFrame()
for i in df.columns[1:]:
  df_final = df.loc[:,['Date',i]]
  df_final['District'] = i 
  df_final.rename(columns={i: 'Value'}, inplace=True)
  df_aux = pd.concat([df_aux, df_final])

df_aux

df_aux.to_csv('../raw/eda-result/rent.csv', index = False, sep = ',')

rent = df_aux

rent['Date'].unique().tolist()

district = ['Centro', 'Arganzuela', 'Barajas']

fig = px.line(rent, x=rent['Date'].unique().tolist(), y=rent['Value'][(rent['District']==district[0])], markers=True)
fig.update_xaxes(tickangle = 270)

for d in district[1:]:

    fig.add_scatter(x=rent['Date'].unique().tolist(), y=rent['Value'][(rent['District']==d)])

fig.show()
