==> notebooks/7-h3.ipynb <==
from keplergl import KeplerGl

# Data manipulation
import pandas as pd # data manipulation and dataframes

import pickle

import json

with open('../data/config_porto_airbnb.txt') as f:
    config = f.readlines()

res = json.loads(config[0])

with open('config_porto_airbnb.pickle', 'wb') as f:
    pickle.dump(res, f)
